# openapi/vaultchain.yaml
openapi: 3.1.0
info:
  title: VaultChain API
  version: 0.1.0
  description: |
    **FICTIONAL** mock digital-asset custody platform — the system-under-test for a
    showcase test architecture. No real blockchain, keys, or money. This spec is the
    contract (source of truth): the P2 contract-test layer independently re-encodes it
    as Zod schemas.

    Auth: `X-Api-Key` header. Roles: `CLIENT` (tenant-bound), `OPERATOR`,
    `COMPLIANCE_OFFICER`, `ADMIN`. Errors are RFC 9457 `application/problem+json`.
    Amounts are decimal strings; `*Minor` fields are integer minor-unit strings
    interpreted with the asset's `decimals`.
servers:
  - url: http://localhost:3000
security:
  - apiKey: []
...
paths:
...
  /simulator/state:
    get:
      tags: [simulator]
      summary: Current chain/clock/fault state (ADMIN)
      responses:
        '200':
          description: Chain state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChainState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /simulator/chain/advance:
    post:
      tags: [simulator]
      summary: Advance block height (and, unless frozen, the sim clock) (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [blocks]
              additionalProperties: false
              properties:
                blocks: { type: integer, minimum: 1, maximum: 10000 }
      responses:
        '200':
          description: Advanced; settled transactions counted
          content:
            application/json:
              schema:
                type: object
                required: [blockHeight, simClockMs, settled]
                properties:
                  blockHeight: { type: integer }
                  simClockMs: { type: string }
                  settled: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /simulator/clock/set:
    post:
      tags: [simulator]
      summary: Set the sim clock (ms) (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ms]
              additionalProperties: false
              properties:
                ms: { type: string, pattern: '^\d+$' }
      responses:
        '200':
          description: Chain state after set
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChainState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /simulator/clock/advance:
    post:
      tags: [simulator]
      summary: Atomically advance the sim clock by a relative amount (ADMIN)
      description: |
        Unlike clock/set (absolute, last-writer-wins), relative advances
        serialize on the store's write lock and can never rewind the clock
        past a concurrent caller's progress — the primitive parallel test
        workers must use.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ms]
              additionalProperties: false
              properties:
                ms: { type: string, pattern: '^\d+$' }
      responses:
        '200':
          description: Chain state after the advance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChainState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /simulator/clock/freeze:
    post:
      tags: [simulator]
      summary: Freeze/unfreeze the sim clock (ADMIN)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                frozen: { type: boolean, default: true }
      responses:
        '200':
          description: Chain state after change
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChainState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /simulator/tx/{id}/force:
    post:
      tags: [simulator]
      summary: Force a transaction outcome; failed debited withdrawals are refunded (ADMIN)
      parameters: [{ $ref: '#/components/parameters/id' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome]
              additionalProperties: false
              properties:
                outcome: { type: string, enum: [CONFIRMED, FAILED] }
      responses:
        '200':
          description: Transaction after forcing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TransactionRaw' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /simulator/screening/next:
    post:
      tags: [simulator]
      summary: Queue the next screening decision (consumed once) (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [outcome]
              additionalProperties: false
              properties:
                outcome: { type: string, enum: [CLEAN, FLAG] }
      responses:
        '200':
          description: Queued
          content:
            application/json:
              schema:
                type: object
                required: [queued]
                properties:
                  queued: { type: string, enum: [CLEAN, FLAG] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /simulator/webhooks/{id}/replay:
    post:
      tags: [simulator]
      summary: Re-deliver a webhook; deposit events re-enter the credit handler (ADMIN)
      parameters: [{ $ref: '#/components/parameters/id' }]
      responses:
        '200':
          description: Replay result (creditResult null for non-deposit events)
          content:
            application/json:
              schema:
                type: object
                required: [delivery, creditResult]
                properties:
                  delivery: { $ref: '#/components/schemas/WebhookDelivery' }
                  creditResult:
                    oneOf:
                      - type: 'null'
                      - type: object
                        required: [credited, reason]
                        properties:
                          credited: { type: boolean }
                          reason: { type: string, enum: [credited, already-credited, not-creditable] }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /simulator/webhooks/delay:
    post:
      tags: [simulator]
      summary: Delay all webhook deliveries by N sim-clock ms (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [ms]
              additionalProperties: false
              properties:
                ms: { type: string, pattern: '^\d+$' }
      responses:
        '200':
          description: Set
          content:
            application/json:
              schema:
                type: object
                required: [webhookDelayMs]
                properties:
                  webhookDelayMs: { type: string }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /simulator/reset:
    post:
      tags: [simulator]
      summary: Reset chain/clock/fault state to defaults — not the database (ADMIN)
      responses:
        '200':
          description: Chain state after reset
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ChainState' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Missing or invalid X-Api-Key
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: Role not permitted
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Not found (including cross-tenant reads)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: State conflict (e.g. approvals closed, duplicate decision)
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 problem details
      required: [type, title, status]
      properties:
        type: { type: string, format: uri }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
...
    TransactionRaw:
      type: object
      description: Transaction as persisted (simulator force response)
      required: [id, walletId, type, assetSymbol, amountMinor, feeMinor, state]
      properties:
        id: { type: string }
        walletId: { type: string }
        type: { type: string, enum: [DEPOSIT, WITHDRAWAL] }
        assetSymbol: { type: string }
        amountMinor: { type: string }
        feeMinor: { type: string }
        state: { type: string }
...
    WebhookDelivery:
      type: object
      required: [id, subscriptionId, event, payload, signature, attempts, status]
      properties:
        id: { type: string }
        subscriptionId: { type: string }
        event: { type: string }
        payload: { type: string, description: Raw JSON string the signature covers }
        signature: { type: string, description: 'sha256=<hex> HMAC over the raw payload' }
        attempts: { type: integer }
        status: { type: string, enum: [PENDING, DELIVERED] }
        dueAtSimMs: { type: string }
        deliveredAt: { type: [string, 'null'], format: date-time }
        createdAt: { type: string, format: date-time }
    ChainState:
      type: object
      required: [id, blockHeight, simClockMs, frozen, webhookDelayMs, screeningNext]
      properties:
        id: { type: string, const: chain }
        blockHeight: { type: integer }
        simClockMs: { type: string }
        frozen: { type: boolean }
        webhookDelayMs: { type: string }
        screeningNext: { type: [string, 'null'], enum: [CLEAN, FLAG, null] }
        forcedOutcomesJson: { type: string }